requirements.txt
numpy==1.23.0
tensorflow==2.12.0
matplotlib==3.6.0
scikit-learn==1.1.1
streamlit==1.10.0
Pillow==9.1.0
distutils
